Moving class into target namespace.
